* Pass VeryBig Tests

+++ src/Lua/CodeAnalysis/Compilation/Descriptions.cs
@@ -5,7 +5,7 @@ namespace Lua.CodeAnalysis.Compilation
 {
     public readonly record struct LocalVariableDescription
     {
-        public required byte RegisterIndex { get; init; }
+        public required ushort RegisterIndex { get; init; }
         public required int StartPc { get; init; }
     }
 
@@ -20,7 +20,7 @@ namespace Lua.CodeAnalysis.Compilation
     {
         public required ReadOnlyMemory<char> Name { get; init; }
         public required int Index { get; init; }
-        public required byte RegisterIndex { get; init; }
+        public required ushort RegisterIndex { get; init; }
     }
 
     public readonly record struct GotoDescription
