+++ src/Lua/Runtime/LocalValueInfo.cs
@@ -3,7 +3,7 @@
 public readonly record struct LocalValueInfo
 {
     public required ReadOnlyMemory<char> Name { get; init; }
-    public required byte Index { get; init; }
+    public required ushort Index { get; init; }
     public required int StartPc { get; init; }
     public required int EndPc { get; init; }
 }
